+++ src/main/java/net/silentchaos512/gems/entity/EnderSlimeEntity.java
@@ -1,6 +1,7 @@
 package net.silentchaos512.gems.entity;
 
 import net.minecraft.entity.*;
+import net.minecraft.entity.monster.MonsterEntity;
 import net.minecraft.entity.monster.SlimeEntity;
 import net.minecraft.fluid.Fluid;
 import net.minecraft.particles.IParticleData;
@@ -19,6 +20,7 @@ import net.minecraftforge.api.distmarker.OnlyIn;
 import net.silentchaos512.gems.SilentGems;
 
 import javax.annotation.Nonnull;
+import java.util.Random;
 
 public class EnderSlimeEntity extends SlimeEntity {
     private static final ResourceLocation LOOT_TABLE = SilentGems.getId("ender_slime");
@@ -43,6 +45,10 @@ public class EnderSlimeEntity extends SlimeEntity {
     @Override
     public boolean canSpawn(IWorld worldIn, SpawnReason spawnReasonIn) {
         return worldIn.getDifficulty() != Difficulty.PEACEFUL;
+    }
+
+    public static boolean canSpawnAt(EntityType<EnderSlimeEntity> type, IWorld world, SpawnReason reason, BlockPos pos, Random random) {
+        return world.getDifficulty() != Difficulty.PEACEFUL && pos.getY() < 65 && MonsterEntity.isValidLightLevel(world, pos, random);
     }
 
     @Override